Fixed a big bug!
People has reported that on the latest python version the software does not work but now fixed and will soon be avaible to download on my website. Just need to test it in my virtual machine to begin with. 
:)
Thanks to people that reported the bug. My other softwares seems to be getting the issue as well. 
Jonathan Steadman. Bug fixed.